Update the base application to hello-react
