--- package.json
@@ -1,23 +1,0 @@
-{
-  "name": "minicart",
-  "version": "1.0.0",
-  "description": "VTEX Minicart Component",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/vtex-apps/minicart.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/vtex-apps/minicart/issues"
-  },
-  "homepage": "https://github.com/vtex-apps/minicart#readme",
-  "devDependencies": {
-    "eslint": "^4.19.1",
-    "eslint-config-google": "^0.9.1"
-  }
-}
