--- react/StoreTemplate.js
@@ -1,21 +1,0 @@
-import React, { Component } from 'react'
-import PropTypes from 'prop-types'
-import Header from './components/Header'
-
-export default class StoreTemplate extends Component {
-  static propTypes = {
-    children: PropTypes.element,
-  }
-
-  render() {
-    return (
-      <div>
-        <Header />
-        <div className="z-1">{this.props.children}</div>
-        <div className="mt4">
-
-        </div>
-      </div>
-    )
-  }
-}
