--- react/__tests__/addToCartMutation.test.js
@@ -1,7 +1,0 @@
-import addToCartMutation from '../mutations/addToCartMutation.gql'
-
-describe('addToCartMutation test', () => {
-  it('should be the correct mutation', () => {
-    expect(addToCartMutation).toMatchSnapshot()
-  })
-})
