--- react/__tests__/orderFormQuery.test.js
@@ -1,7 +1,0 @@
-import orderFormQuery from '../queries/orderFormQuery.gql'
-
-describe('orderFormQuery test', () => {
-  it('should be the correct query', () => {
-    expect(orderFormQuery).toMatchSnapshot()
-  })
-})
