--- react/__tests__/productQuery.test.js
@@ -1,7 +1,0 @@
-import productQuery from '../queries/productQuery.gql'
-
-describe('productQuery test', () => {
-  it('should be the correct query', () => {
-    expect(productQuery).toMatchSnapshot()
-  })
-})
