--- react/__tests__/productsQuery.test.js
@@ -1,7 +1,0 @@
-import productsQuery from '../queries/productsQuery.gql'
-
-describe('productsQuery test', () => {
-  it('should be the correct query', () => {
-    expect(productsQuery).toMatchSnapshot()
-  })
-})
