--- react/components/minicart.css
@@ -1,4 +1,0 @@
-.minicart {
-    top: 100%;
-    width: 22rem;
-}
