+++ react/index.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function HelloWorld () {
+  return (
+    <div>
+      <h1>Hello world!</h1>
+    </div>
+  )
+}
