--- react/locales/en-US.json
@@ -1,8 +1,0 @@
-{
-  "dreamstore.soon": "Soon...",
-  "dreamstore.search": "Search",
-  "dreamstore.search-placeholder": "Search for products, brands...",
-  "dreamstore.cart": "Cart",
-  "dreamstore.buy": "Buy",
-  "dreamstore.poweredBy": "POWERED BY VTEX"
-}
