--- react/locales/es-AR.json
@@ -1,8 +1,0 @@
-{
-  "dreamstore.soon": "Breve...",
-  "dreamstore.search": "Buscar",
-  "dreamstore.search-placeholder": "Búsqueda por productos, marcas...",
-  "dreamstore.cart": "Carrito",
-  "dreamstore.buy": "Comprar",
-  "dreamstore.poweredBy": "IMPULSADO POR VTEX"
-}
