--- react/locales/pt-BR.json
@@ -1,8 +1,0 @@
-{
-  "dreamstore.soon": "Em breve...",
-  "dreamstore.search": "Buscar",
-  "dreamstore.search-placeholder": "Busque por produtos, marcas...",
-  "dreamstore.cart": "Carrinho",
-  "dreamstore.buy": "Comprar",
-  "dreamstore.poweredBy": "FORNECIDO POR VTEX"
-}
