--- react/mutations/addToCartMutation.gql
@@ -1,8 +1,0 @@
-mutation addItem($orderFormId: String, $items: [OrderFormItemInput])
-  @context(scope: "private") {
-  addItem(orderFormId: $orderFormId, items: $items) {
-    items {
-      id
-    }
-  }
-}
