--- react/mutations/updateItemsMutation.gql
@@ -1,6 +1,0 @@
-mutation updateItems($orderFormId: String, $items: [OrderFormItemInput])
-  @context(scope: "private") {
-  updateItems(orderFormId: $orderFormId, items: $items) {
-    orderFormId
-  }
-}
