--- react/queries/orderFormQuery.gql
@@ -1,15 +1,0 @@
-query orderForm @context(scope: "private") {
-  orderForm {
-    orderFormId
-    value
-    items {
-      id
-      name
-      imageUrl
-      detailUrl
-      skuName
-      quantity
-      sellingPrice
-    }
-  }
-}
