--- react/queries/productQuery.gql
@@ -1,39 +1,0 @@
-query Product($slug: String) {
-  product(slug: $slug) {
-    productName
-    productId
-    description
-    items {
-      itemId
-      name
-      nameComplete
-      complementName
-      ean
-      measurementUnit
-      unitMultiplier
-      images {
-        imageId
-        imageLabel
-        imageTag
-        imageUrl
-        imageText
-      }
-      sellers {
-        sellerId
-        sellerName
-        addToCartLink
-        sellerDefault
-        commertialOffer {
-          Price
-          ListPrice
-          PriceWithoutDiscount
-          RewardValue
-          PriceValidUntil
-          AvailableQuantity
-          Tax
-          CacheVersionUsedToCallCheckout
-        }
-      }
-    }
-  }
-}
