--- react/queries/productsQuery.gql
@@ -1,60 +1,0 @@
-query Products(
-    $query: String
-    $category: String
-    $specificationFilters: [String]
-    $priceRange: String
-    $collection: String
-    $orderBy: String
-    $from: Int
-    $to: Int
-    $salesChannel: String
-  ) {
-    products(
-      query: $query
-      category: $category
-      specificationFilters: $specificationFilters
-      priceRange: $priceRange
-      collection: $collection
-      orderBy: $orderBy
-      from: $from
-      to: $to
-      salesChannel: $salesChannel
-    ) {
-      productId
-      productName
-      linkText
-      description
-      items {
-        itemId
-        name
-        nameComplete
-        complementName
-        ean
-        referenceId {
-          Key
-          Value
-        }
-        measurementUnit
-        unitMultiplier
-        images {
-          imageId
-          imageLabel
-          imageTag
-          imageUrl
-          imageText
-        }
-        sellers {
-          sellerId
-          sellerName
-          commertialOffer {
-            Price
-            ListPrice
-            PriceWithoutDiscount
-            RewardValue
-            PriceValidUntil
-            AvailableQuantity
-          }
-        }
-      }
-    }
-  }
