--- react/theme.css
@@ -1,28 +1,0 @@
-@import url(//io.vtex.com.br/fonts/fabriga/stylesheet.css);
-@import url("./slider.css");
-
-body,
-button,
-input,
-optgroup,
-select,
-textarea,
-span,
-h1,
-h2,
-h3,
-h4,
-p1 {
-  font-family: Fabriga, -apple-system, BlinkMacSystemFont, avenir next, avenir,
-    helvetica neue, helvetica, ubuntu, roboto, noto, segoe ui, arial, sans-serif;
-}
-
-body {
-  background-color: white;
-}
-
-img {
-  max-width: 100%;
-  height: auto;
-  width: auto; /* ie8 */
-}
